Run in the axon-ci container. Restructure all tests that previously started their own Docker containers.
