+++ azure-pipelines.yml
@@ -6,13 +6,17 @@ trigger:
     include:
       - '*'
 
+resources:
+  containers:
+  - container: axon-ci
+    image: wpilib/axon-ci:latest
+
 stages:
 - stage: Build
   jobs:
     - job: Linux
 
-      pool:
-        vmImage: ubuntu-16.04
+      container: axon-ci
 
       steps:
         - task: DockerInstaller@0
@@ -27,7 +31,7 @@ stages:
             workingDirectory: ''
             gradleWrapperFile: 'gradlew'
             gradleOptions: '-Xmx3072m'
-            options: '--stacktrace -PlogTests -Pheadless -PhasDockerSupport'
+            options: '--stacktrace -PlogTests -Pheadless -PhasDockerSupport -PhasTensorFlowSupport'
             javaHomeOption: 'JDKVersion'
             jdkVersionOption: '1.11'
             jdkArchitectureOption: 'x64'
@@ -36,7 +40,7 @@ stages:
             tasks: 'check'
 
         - task: Gradle@2
-          displayName: 'Check'
+          displayName: 'JacocoRootReport'
           inputs:
             workingDirectory: ''
             gradleWrapperFile: 'gradlew'
@@ -57,8 +61,7 @@ stages:
 
     - job: MacOS
 
-      pool:
-        vmImage: macOS-10.14
+      container: axon-ci
 
       steps:
         - task: Gradle@2
@@ -75,7 +78,7 @@ stages:
             tasks: 'check'
 
         - task: Gradle@2
-          displayName: 'Check'
+          displayName: 'JacocoRootReport'
           inputs:
             workingDirectory: ''
             gradleWrapperFile: 'gradlew'
@@ -90,8 +93,7 @@ stages:
 
     - job: Windows
 
-      pool:
-        vmImage: windows-2019
+      container: axon-ci
 
       steps:
         - task: Gradle@2
@@ -108,7 +110,7 @@ stages:
             tasks: 'check'
 
         - task: Gradle@2
-          displayName: 'Check'
+          displayName: 'JacocoRootReport'
           inputs:
             workingDirectory: ''
             gradleWrapperFile: 'gradlew'
