+++ training-test-util/src/main/kotlin/edu/wpi/axon/training/testutil/TrainTestUtil.kt
@@ -8,10 +8,8 @@ import io.kotlintest.matchers.file.shouldExist
 import io.kotlintest.matchers.string.shouldNotBeEmpty
 import io.kotlintest.shouldBe
 import java.io.File
+import java.nio.file.Files
 import java.nio.file.Paths
-import mu.KotlinLogging
-
-private val LOGGER = KotlinLogging.logger("training-test-util")
 
 /**
  * Loads a model with name [modelName] from the test resources.
@@ -29,42 +27,29 @@ fun loadModel(modelName: String, stub: () -> Unit): Pair<Model, String> {
 }
 
 /**
- * Tests that a training scripts works by running it in the axon-ci Docker container and asserting
- * that the expected trained model file is written to disk.
+ * Tests that a training scripts works by running it and asserting that the trained model file was
+ * written to disk.
  *
- * @param oldModelPath The path to load the old model from.
- * @param oldModelName The name of the old model.
- * @param newModelName The name of the new model.
+ * @param dir The working directory to run the script in.
  * @param script The content of the script to run.
- * @param dir The working directory.
+ * @param newModelName The name of the new model.
  */
 fun testTrainingScript(
-    oldModelPath: String,
-    oldModelName: String,
-    newModelName: String,
-    script: String,
     dir: File,
-    cleanupAfterCommand: String? = null
+    script: String,
+    newModelName: String
 ) {
-    Paths.get(oldModelPath).toFile().copyTo(Paths.get(dir.absolutePath, oldModelName).toFile())
-    Paths.get(dir.absolutePath, "script.py").toFile().writeText(script)
-
+    val scriptFile = Files.createTempFile(dir.toPath(), "", ".py").toFile()
+    scriptFile.writeText(script)
     runCommand(
         listOf(
-            "docker",
-            "run",
-            "--rm",
-            "-v",
-            "${dir.absolutePath}:/home",
-            "wpilib/axon-ci:latest",
-            cleanupAfterCommand?.let {
-                "/usr/bin/python3.6 /home/script.py && $it"
-            } ?: "/usr/bin/python3.6 /home/script.py"
+            "python3.6",
+            scriptFile.path
         ),
         emptyMap(),
         dir
     ).attempt().unsafeRunSync().shouldBeRight { (exitCode, stdOut, stdErr) ->
-        LOGGER.info {
+        println(
             """
             |Process std out:
             |$stdOut
@@ -73,10 +58,10 @@ fun testTrainingScript(
             |$stdErr
             |
             """.trimMargin()
-        }
+        )
 
         stdOut.shouldNotBeEmpty()
         exitCode shouldBe 0
-        Paths.get(dir.absolutePath, newModelName).shouldExist()
+        Paths.get(newModelName).shouldExist()
     }
 }
